test: put test class into right package
