+++ src/test/java/racingCar/domain/RacingCarsTest.java
@@ -1,4 +1,4 @@
-package racingCar;
+package racingCar.domain;
 
 import static org.assertj.core.api.Assertions.*;
