+++ src/test/java/racingCar/validator/RacingTimesTest.java
@@ -1,4 +1,4 @@
-package racingCar;
+package racingCar.validator;
 
 import static org.assertj.core.api.Assertions.*;
